tried to reduce requests
Tried to reduce the amount of requests (made the program much slower). Plan is: get the ID for each company (dont get ID from duplicates) then make dictionary with ID and the name of the company.

Now when we want a name we can simply get the ID which is much easier to access (no requests) and then use the ID to access the name.